==> R/00_all.qmd <==
---
title: "00_all"
format: 
  html:
    embed-resources: true
editor: visual
---

```{r}
#| warning: false
#| error: false
#| message: true
library(quarto)
library(knitr)
source("99_proj_func.R")
results_path <-  "../results/"
r_path <-  "../R/"

#Long runtime, 20-30 minutes
knit(str_c(r_path,"01_load.qmd"), output = tempfile(fileext = ".md"))
knit(str_c(r_path,"02_clean.qmd"), output = tempfile(fileext = ".md"))
knit(str_c(r_path,"03_augment.qmd"), output = tempfile(fileext = ".md"))
knit(str_c(r_path,"04_describe.qmd"), output = tempfile(fileext = ".md"))
knit(str_c(r_path,"05_analysis.qmd"), output = tempfile(fileext = ".md"))
knit(str_c(r_path,"06_analysis.qmd"), output = tempfile(fileext = ".md"))
```

```{r}
quarto_render(input = str_c(r_path, "01_load.qmd"),
  quarto_args = c("--output-dir", results_path))

quarto_render(input = str_c(r_path, "02_clean.qmd"),
  quarto_args = c("--output-dir", results_path))

quarto_render(input = str_c(r_path, "03_augment.qmd"),
  quarto_args = c("--output-dir", results_path))

quarto_render(input = str_c(r_path, "04_describe.qmd"),
  quarto_args = c("--output-dir", results_path))

quarto_render(input = str_c(r_path, "05_analysis.qmd"),
  quarto_args = c("--output-dir", results_path))

quarto_render(input = str_c(r_path, "06_analysis.qmd"),
  quarto_args = c("--output-dir", results_path))
```
