==> R/01_load.qmd <==
---
title: "01_load.qmd"
format: 
  html:
    embed-resources: true
editor: visual
---

## Loading libraries and functions

```{r}
#| warning: false
#| error: false
#| message: false
library(tidyverse)
library(purrr)
library(readr)
library(dplyr)
library(stringr)
library(curl)
source("99_proj_func.R")
```

## Downloading SIDER data

```{r}
#| warning: false
#| error: false
#| message: false
# SIDER base URL
base_url <- "http://sideeffects.embl.de/media/download/"
data_path = "../data/"
raw_path = str_c(data_path, "_raw/")
sider_path = str_c(raw_path, "sider/")
drugbank_path = str_c(raw_path, "drugbank/")

# Create directory for SIDER data
dir.create(sider_path, recursive = TRUE, showWarnings = FALSE)

# Files to download
files <- c(
  "meddra_all_se.tsv.gz",
  "meddra_freq.tsv.gz",
  "meddra_all_label_se.tsv.gz",
  "meddra_all_indications.tsv.gz",
  "drug_names.tsv",
  "drug_atc.tsv"
)

# Downloading
files |> walk(~
  curl_download(
    str_c(base_url, .x),
    file.path(sider_path, .x),
    quiet = TRUE
  )
)
```

## Adding column names to SIDER dataset

```{r}
#| warning: false
#| error: false
#| message: false

# Load side effects (meddra_all_se)
sider_se <- read_tsv(
  str_c(sider_path,"meddra_all_se.tsv.gz"),
  col_names = c("STITCH_flat", "STITCH_stereo",
                "UMLS_CUI_label", "MedDRA_concept_type",
                "UMLS_MedDRA", "side_effect"),
  col_types = cols()
)

# Load frequency data (meddra_freq)
sider_freq <- read_tsv(
  str_c(sider_path,"meddra_freq.tsv.gz"),
  col_names = c("STITCH_flat", "STITCH_stereo",
                "UMLS_CUI_label", "placebo",
                "freq_description", "freq_lower", "freq_upper",
                "MedDRA_concept_type", "UMLS_MedDRA", "side_effect"),
  col_types = cols()
)

# Load indications (meddra_all_indications)
sider_indications <- read_tsv(
  str_c(sider_path,"meddra_all_indications.tsv.gz"),
  col_names = c("STITCH_flat", "UMLS_CUI_label",
               "method_of_detection", "concept_name",
               "MedDRA_concept_type", "UMLS_MedDRA",
               "MedDRA_concept_name"),
  col_types = cols()
)

# Load drug names
drug_names <- read_tsv(
  str_c(sider_path,"drug_names.tsv"),
  col_names = c("STITCH_flat", "drug_name"),
  col_types = cols()
)


# Load ATC codes
drug_atc <- read_tsv(
  str_c(sider_path,"drug_atc.tsv"),
  col_names = c("STITCH_flat", "ATC_code"),
  col_types = cols()
)

```

## Saving SIDER data

```{r}
write_tsv(sider_se, str_c(data_path,"sider_se.tsv"))
write_tsv(sider_freq, str_c(data_path,"sider_freq.tsv"))
write_tsv(sider_indications, str_c(data_path,"sider_indications.tsv"))
write_tsv(drug_names, str_c(data_path,"drug_names.tsv"))
write_tsv(drug_atc, str_c(data_path,"drug_atc.tsv"))
```

## Downloading the Drugbank data set

```{r}
#| warning: false
#| error: false
#| message: false
# Make directory
dir.create(drugbank_path, recursive = TRUE, showWarnings = FALSE)

# Files to download
urls <- c(
  vocabulary = 
    "https://go.drugbank.com/releases/5-1-13/downloads/all-drugbank-vocabulary",
  structures = 
    "https://go.drugbank.com/releases/5-1-13/downloads/all-structure-links",
  targets    = 
    "https://go.drugbank.com/releases/5-1-13/downloads/target-all-uniprot-links"
)

zip_paths <- file.path(drugbank_path, str_c("drugbank_", names(urls), ".zip"))

# Credentials
user <- "s243161@dtu.dk"
password <- "&5O&w6inYYPc^"           

# Download
walk2(
  urls,
  zip_paths,
  ~ curl_download(
      url = .x,
      destfile = .y,
      quiet = TRUE,
      handle = new_handle(userpwd = str_c(user, ":", password))
    )
)

# Unzip
walk(zip_paths, ~ unzip(.x, exdir = drugbank_path))

# Delete zips
zip_paths |>
  file.remove() |>
  invisible()

vocabulary <- read_csv(str_c(drugbank_path,"drugbank vocabulary.csv"))
structures <- read_csv(str_c(drugbank_path,"structure links.csv"))
targets <- read_csv(str_c(drugbank_path,"uniprot links.csv"))
```

## Saving Drugbank datasets

```{r}
write_tsv(vocabulary, str_c(data_path,"drugbank_vocabulary.tsv"))
write_tsv(structures, str_c(data_path,"drugbank_structures.tsv"))
write_tsv(targets, str_c(data_path,"drugbank_targets.tsv"))
```

### Check if results folder exists otherwise create it

```{r}
if (!dir.exists("../results")) {
  dir.create("../results")
}
```
