==> R/03_augment.qmd <==
---
title: "03_augment.qmd"
format: 
  html:
    embed-resources: true
editor: visual
---

## Loading libraries

```{r}
#| message: false
library(tidyverse)
library(dplyr)
library(stringr)
library(rJava)
library(rcdk)
library(RChemMass)
library(rJava)
library(Rcpi)
source("99_proj_func.R")
```

## Loading cleaned, full data

```{r}
#| message: false
data_path = "../data/"
sider = read_tsv(str_c(data_path, "sider_full_clean.tsv"))
drugbank = read_tsv(str_c(data_path, "drugbank_full_clean.tsv"))
```

## Adding Lipinski features and fixing column names

```{r}
#| message: false
#| warning: false
sider <- sider |>
  rename("common_name"=drug_name)
drugbank <- drugbank |>
  mutate(common_name = str_to_lower(common_name)) |>
  group_by(drugbank_id) |>
  summarise(
    across(
      .cols = -uniprot_name,
      .fns = ~ first(.x)
    ),
    uniprot_names = list(unique(uniprot_name)),
    .groups = "drop"
  )
# This takes quite some time (10-20min), but it works
drugbank <- drugbank |>
  drop_na(smiles) |>
  mutate(molecular_weight = map_dbl(smiles, smiles2mass)) |>
  mutate(hbd = map_dbl(smiles, count_hbd)) |>
  mutate(hba = map_dbl(smiles, count_hba)) |>
  mutate(clog_p = map_dbl(smiles, calc_clogp)) |>
  mutate(lipinski = hbd <= 5 &
                    hba <= 10 &
                    molecular_weight < 500 &
                    clog_p <= 5)

```

```{r}
df <- drugbank |> inner_join(sider, by="common_name")
```

## Export as tsv

```{r}
#some issues with writing the uniprot names because of list format. this should fix. load with splitting by ";"
df <- df |>
  mutate(
    uniprot_names = map_chr(uniprot_names, ~ paste(.x, collapse = ";"))
  )
write_tsv(df, str_c(data_path, "merged_data.tsv"))
```
