==> R/06_analysis.qmd <==
---
title: "06_analysis_2"
format: 
  html:
    embed-resources: true
editor: visual
---

## Loading libraries and functions

```{r}
#| message: false
library(tidyverse)
source("99_proj_func.R")
```

## Loading dataset

```{r}
#| message: false
#| warning: false
data_path = "../data/"
results_path = "../results/"
presentation_path = "..doc/presentation_files/figures/"
df <- read_tsv(str_c(data_path, "merged_data.tsv"))
```

## Combine all side-effect for each drug into a single row

```{r}
df_drug <- df |>
  group_by(drugbank_id) |>
  summarise(
    common_name = first(common_name),
    drug_groups = first(drug_groups),
    smiles = first(smiles),
    formula = first(formula),
    uniprot_names = first(uniprot_names),
    molecular_weight = first(molecular_weight),
    hbd = first(hbd),
    hba = first(hba),
    clog_p = first(clog_p),
    lipinski = first(lipinski),
    # side effects
    side_effects = str_c(unique(side_effect_PT), collapse = "; "),
    side_effect_burden = n_distinct(side_effect_PT),
    .groups = "drop"
  )
```

## Distributions of Lipinski-related descriptors

```{r}
# Histogram for H-bond donors (HBD) 
hbd_hist <- df_drug |>
  ggplot(aes(x = hbd,
             y = after_stat(count / sum(count) * 100))) +
  geom_histogram(binwidth = 1,
                 fill = "steelblue",
                 color = "black") +
  theme_bw() +
  labs(x = "H-bond donors (HBD)",
       y = "Percentage (%)")

ggsave(str_c(results_path,"Analysis2_HBD_count.png"), plot = hbd_hist, 
       width = 8, height = 6, dpi = 300)
ggsave(str_c(presentation_path,"Analysis2_HBD_count.png"), plot = hbd_hist, 
       width = 8, height = 6, dpi = 300)
print(hbd_hist)
# Histogram for H-bond acceptors (HBA) 
hba_hist <- df_drug |>
  ggplot(aes(x = hba,
             y = after_stat(count / sum(count) * 100))) +
  geom_histogram(binwidth = 1,
                 fill = "steelblue",
                 color = "black") +
  theme_bw() +
  labs(x = "H-bond acceptors (HBA)",
       y = "Percentage (%)")

ggsave(str_c(results_path,"Analysis2_HBA_count.png"), plot = hba_hist, 
       width = 8, height = 6, dpi = 300)
ggsave(str_c(presentation_path,"Analysis2_HBA_count.png"), plot = hba_hist, 
       width = 8, height = 6, dpi = 300)
print(hba_hist)
# Bar plot for Lipinski rule compliance as percentage
lipinski_bar <- df_drug |>
  ggplot(aes(x = lipinski,
             y = after_stat(count / sum(count) * 100))) +
  geom_bar(fill = "steelblue",
           color = "black") +
  theme_bw() +
  labs(x = "Lipinski rule satisfied",
       y = "Percentage (%)")

ggsave(str_c(results_path,"Analysis2_Lipinski_percent.png"), plot = lipinski_bar, 
       width = 8, height = 6, dpi = 300)

print(lipinski_bar)
# Density plot for Molecular Weight (MW)
mw_density <- df_drug |>
  ggplot(aes(x = molecular_weight)) +
  geom_density(fill = "lightgreen",
               alpha = 0.6) +
  theme_bw() +
  labs(x = "Molecular Weight (MW)",
       y = "Density")
ggsave(str_c(results_path,"Analysis2_MW_density.png"), plot = mw_density, 
       width = 8, height = 6, dpi = 300)
ggsave(str_c(presentation_path,"Analysis2_MW_density.png"), plot = mw_density, 
       width = 8, height = 6, dpi = 300)
print(mw_density)
# Density plot for LogP (ClogP)
logp_density <- df_drug |>
  ggplot(aes(x = clog_p)) +
  geom_density(fill = "lightgreen",
               alpha = 0.6) +
  theme_bw() +
  labs(x = "LogP (ClogP)",
       y = "Density")
ggsave(str_c(results_path,"Analysis2_LogP_density.png"), plot = logp_density, 
       width = 8, height = 6, dpi = 300)
ggsave(str_c(presentation_path,"Analysis2_LogP_density.png"), plot = logp_density, 
       width = 8, height = 6, dpi = 300)
print(logp_density)
```

## Correlation plots with side-effect burden

```{r}
# HBD vs side-effect burden
# HBD vs side-effect burden
cor_hbd <- cor(df_drug$hbd, df_drug$side_effect_burden, use = "complete.obs")
hbd_burden <- df_drug |>
  ggplot(aes(x = hbd, y = side_effect_burden)) +
  geom_point(color = "steelblue", alpha = 0.6) +
  geom_smooth(method = "lm", color = "red") +
  labs(
    x = "H-bond donors (HBD)",
    y = "Number of Side Effects",
    title = "HBD vs Side-Effect Burden",
    subtitle = str_c("Correlation coefficient:", round(cor_hbd, 3))
  ) +
  theme_bw()
ggsave(str_c(results_path,"Analysis2_HBD_SE.png"), plot = hbd_burden,
       width = 8, height = 6, dpi = 300)
ggsave(str_c(presentation_path,"Analysis2_HBD_SE.png"), plot = hbd_burden, 
       width = 8, height = 6, dpi = 300)
print(hbd_burden)


# HBA vs side-effect burden
# HBA vs side-effect burden
cor_hba <- cor(df_drug$hba, df_drug$side_effect_burden, use = "complete.obs")
hba_burden <- df_drug |>
  ggplot(aes(x = hba, y = side_effect_burden)) +
  geom_point(color = "steelblue", alpha = 0.6) +
  geom_smooth(method = "lm", color = "red") +
  labs(
    x = "H-bond acceptors (HBA)",
    y = "Number of Side Effects",
    title = "HBA vs Side-Effect Burden",
    subtitle = str_c("Correlation coefficient:", round(cor_hba, 3))
  ) +
  theme_bw()
ggsave(str_c(results_path,"Analysis2_HBA_SE.png"), plot = hba_burden, 
       width = 8, height = 6, dpi = 300)
ggsave(str_c(presentation_path,"Analysis2_HBA_SE.png"), plot = hba_burden, 
       width = 8, height = 6, dpi = 300)
print(hba_burden)

# Molecular Weight vs side-effect burden
cor_mw <- cor(df_drug$molecular_weight, df_drug$side_effect_burden, use = "complete.obs")
mw_burden <- df_drug |>
  ggplot(aes(x = molecular_weight, y = side_effect_burden)) +
  geom_point(color = "steelblue", alpha = 0.6) +
  geom_smooth(method = "lm", color = "red") +
  labs(
    x = "Molecular Weight (MW)",
    y = "Number of Side Effects",
    title = "MW vs Side-Effect Burden",
    subtitle = str_c("Correlation coefficient:", round(cor_mw, 3))
  ) +
  theme_bw()
ggsave(str_c(results_path,"Analysis2_MW_SE.png"), plot = mw_burden,
       width = 8, height = 6, dpi = 300)
ggsave(str_c(presentation_path,"Analysis2_MW_SE.png"), plot = mw_burden,
       width = 8, height = 6, dpi = 300)
print(mw_burden)

# LogP vs side-effect burden
cor_logp <- cor(df_drug$clog_p, df_drug$side_effect_burden, use = "complete.obs")
logp_burden <- df_drug |>
  ggplot(aes(x = clog_p, y = side_effect_burden)) +
  geom_point(color = "steelblue", alpha = 0.6) +
  geom_smooth(method = "lm", color = "red") +
  labs(
    x = "LogP (ClogP)",
    y = "Number of Side Effects",
    title = "LogP vs Side-Effect Burden",
    subtitle = str_c("Correlation coefficient:", round(cor_logp, 3))
  ) +
  theme_bw()
ggsave(str_c(results_path,"Analysis2_LogP_SE.png"),
       plot = logp_burden, width = 8, height = 6, dpi = 300)
ggsave(str_c(presentation_path,"Analysis2_LogP_SE.png"), 
       plot = logp_burden, width = 8, height = 6, dpi = 300)
print(logp_burden)

# Lipinski rule vs side-effect burden
lipinkski_burden <- df_drug |>
  ggplot(aes(x = lipinski, y = side_effect_burden)) +
  geom_boxplot(fill = "steelblue", color = "black") +
  theme_bw() +
  labs(
    x = "Lipinski Rule Satisfied",
    y = "Number of Side Effects",
    title = "Lipinski Rule vs Side-Effect Burden",
  )
ggsave(str_c(results_path,"Analysis2_Lipinski_SE.png"), plot = lipinkski_burden,
       width = 8, height = 6, dpi = 300)
ggsave(str_c(presentation_path,"Analysis2_Lipinski_SE.png"), plot = lipinkski_burden, 
       width = 8, height = 6, dpi = 300)
print(lipinkski_burden)

```

## Legality state

```{r}
df_legal <- df_drug |>
  separate_rows(drug_groups, sep = ";") |> 
  mutate(drug_groups = trimws(drug_groups))     

df_counts <- df_legal |>
  group_by(drug_groups, lipinski) |>
  summarise(n_drugs = n_distinct(common_name), .groups = "drop")


legal_plot <- ggplot(df_counts,
       aes(x = drug_groups,
           y = n_drugs,
           fill = lipinski)) +
  geom_bar(stat = "identity",
           color = "black",
           position = position_dodge()) +
  labs(x = "Drug Group",
       y = "Number of Unique Drugs",
       fill = "Lipinski Rule") +
  theme_bw() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))
ggsave(str_c(results_path,"Analysis2_Lipinski_Legal.png"), plot = legal_plot, 
       width = 8, height = 6, dpi = 300)
ggsave(str_c(presentation_path,"Analysis2_Lipinski_Legal.png"), plot = legal_plot, 
       width = 8, height = 6, dpi = 300)
print(legal_plot)
```
